Remove ignoreFolders option from FileSize rule
### What's done:
* Removed code
* Fixed docs
