+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/utils/StringUtils.kt
@@ -59,16 +59,6 @@ fun String.containsOneLetterOrZero(): Boolean {
 }
 
 /**
- * Splits [this] string by file path separator
- *
- * @return list of path parts
- */
-fun String.splitPathToDirs(): List<String> =
-        this.replace("\\", "/")
-            .replace("//", "/")
-            .split("/")
-
-/**
  * method checks that string has prefix like:
  * mFunction, kLength or M_VAR
  *
@@ -105,17 +95,3 @@ fun String.removePrefix(): String {
     }
     return this
 }
-
-/**
- * Checks if [this] String is a name of a kotlin script file by checking whether file extension equals 'kts'
- *
- * @return true if this is a kotlin script file name, false otherwise
- */
-fun String.isKotlinScript() = endsWith(".kts")
-
-/**
- * Checks if [this] String is a name of a gradle kotlin script file by checking whether file extension equals 'gradle.kts'
- *
- * @return true if this is a gradle kotlin script file name, false otherwise
- */
-fun String.isGradleScript() = endsWith("gradle.kts")
