+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/ruleset/chapter3/FileSizeWarnTest.kt
@@ -24,19 +24,11 @@ class FileSizeWarnTest : LintTestBase(::FileSize) {
     )
     private val rulesConfigListIgnore: List<RulesConfig> = listOf(
         RulesConfig(Warnings.FILE_IS_TOO_LONG.name, true,
-            mapOf("maxSize" to "5", "ignoreFolders" to "main"))
+            mapOf("maxSize" to "5"))
     )
     private val rulesConfigListEmpty: List<RulesConfig> = listOf(
         RulesConfig(Warnings.FILE_IS_TOO_LONG.name, true,
             emptyMap())
-    )
-    private val rulesConfigListOnlyIgnore: List<RulesConfig> = listOf(
-        RulesConfig(Warnings.FILE_IS_TOO_LONG.name, true,
-            mapOf("ignoreFolders" to "A"))
-    )
-    private val rulesConfigListTwoIgnoreFolders: List<RulesConfig> = listOf(
-        RulesConfig(Warnings.FILE_IS_TOO_LONG.name, true,
-            mapOf("maxSize" to "8", "ignoreFolders" to "A, B"))
     )
 
     @Test
@@ -91,42 +83,10 @@ class FileSizeWarnTest : LintTestBase(::FileSize) {
             fileName = file.absolutePath, rulesConfigList = rulesConfigListLarge)
     }
 
-    @Test
-    @Tag(WarningNames.FILE_IS_TOO_LONG)
-    fun `config has only ignoreFolders`() {
-        val path = javaClass.classLoader.getResource("test/paragraph3/src/main/A/FileSize2000.kt")
-        val file = File(path!!.file)
-        lintMethod(file.readText(), fileName = file.absolutePath, rulesConfigList = rulesConfigListOnlyIgnore)
-    }
-
     private fun generate2000lines(): Int {
         val path = javaClass.classLoader.getResource("test/paragraph3/src/main/A/FileSize2000.kt")
         val file = File(path!!.file)
         file.writeText("//hello \n".repeat(2000))
         return 2000
-    }
-
-    @Test
-    @Tag(WarningNames.FILE_IS_TOO_LONG)
-    @Suppress("SAY_NO_TO_VAR")
-    fun `ignoring two out of three folders`() {
-        var path = javaClass.classLoader.getResource("test/paragraph3/src/main/A/FileSizeA.kt")
-        var file = File(path!!.file)
-        lintMethod(file.readText(), fileName = file.absolutePath, rulesConfigList = rulesConfigListTwoIgnoreFolders)
-        path = javaClass.classLoader.getResource("test/paragraph3/src/main/B/FileSizeB.kt")
-        file = File(path!!.file)
-        lintMethod(file.readText(), fileName = file.absolutePath, rulesConfigList = rulesConfigListTwoIgnoreFolders)
-        path = javaClass.classLoader.getResource("test/paragraph3/src/main/C/FileSizeC.kt")
-        file = File(path!!.file)
-        val size = SIZE
-        lintMethod(file.readText(),
-            LintError(1, 1, "$DIKTAT_RULE_SET_ID:file-size",
-                "${Warnings.FILE_IS_TOO_LONG.warnText()} $size", false),
-            fileName = file.absolutePath,
-            rulesConfigList = rulesConfigListTwoIgnoreFolders)
-    }
-
-    companion object {
-        private const val SIZE = 10
     }
 }
