+++ info/available-rules.md
@@ -53,7 +53,7 @@
 |    2  |     2.4.1  |  FIRST_COMMENT_NO_BLANK_LINE                  |  Check: warns if there is a new line before first comment<br>Fix: deletes a new line  |  yes  |  no  |  - |
 |    2  |     2.4.1  |  IF_ELSE_COMMENTS                         |  Check: warns if there is a comment not in the else block<br>Fix: adds the comment to the first line in else block  |  yes  |  no  |  - |
 |    2  |     2.2.1  |  HEADER_NOT_BEFORE_PACKAGE                |  Check: warns if header KDoc if file is located not before package directive<br>Fix: moves this KDoc                                                                           |  yes  |  no  | |
-|    3  |     3.1.1  |  FILE_IS_TOO_LONG                         |  Check: warns if file has too many lines<br>Fix: no                                                                                                                            |   no  |  maxSize<br> ignoreFolders  |                                                                                      - |
+|    3  |     3.1.1  |  FILE_IS_TOO_LONG                         |  Check: warns if file has too many lines<br>Fix: no                                                                                                                            |   no  |  maxSize  |                                                                                      - |
 |    3  |     3.1.2  |  FILE_CONTAINS_ONLY_COMMENTS              |  Check: warns if file contains only comments, imports and package directive.                                                                                                   |   no  |  no  |                                                                                      - |
 |    3  |     3.1.2  |  FILE_INCORRECT_BLOCKS_ORDER              |  Check: warns if general order of code parts is wrong.<br>Fix: rearranges them.                                                                                                |  yes  |  no  |  handle other elements that could be present before package directive (other comments) |
 |    3  |     3.1.2  |  FILE_NO_BLANK_LINE_BETWEEN_BLOCKS        |  Check: warns if there is not exactly one blank line between code parts.<br>Fix: leaves single empty line                                                                      |  yes  |  no  |                                                                                      - |
